Extract terminal
